feature(format-manager): add FormatManager class and have it take over old function
Usage:
    var formatManager = new FormatManager();
    formatManager.applyFormats(tFormats, dataTable);

Where tFormats is the definitions (templates) for the formatters, and dataTable is an instance of google.visualization.DataTable.
Returns an object like this: { requirestHtml: boolean }, indicating whether any formatters require the allowHtml option to be set on the chart.

BREAKING CHANGE: Breaks custom formatters. Will implement later.
angular-google-chart/angular-google-chart#177