feat(format-manager): implement custom formatters
BREAKING CHANGE: If there is a google visualization formatter whose naming convention matches the custom format, it will use the google one instead of the custom formatter.  This could be reversed, but I felt that google's formatter should be the first choice. Otherwise follows the behaviour described by the original author of this feature.  angular-google-chart/angular-google-chart#49 angular-google-chart/angular-google-chart#177