test: added tests for googleChartApiPromise service
Tests to be sure the service returned a promise by checking for the standard promise methods. Also tests that the promise resolves to the global google object. This second part makes this more of an end-to-end test than a unit test. Maybe this could be avoided by mocking the document and faking the script tag's ready event, etc. I just don't have time to deep dive into this suite of tests right now.